+++ src/js/api/instance.js
@@ -13,12 +13,12 @@ export default function (UIkit) {
         }
 
         if (el[DATA][name]) {
-            console.warn(`Component "${name}" is already mounted on element: `, el);
             return;
         }
 
         el[DATA][name] = this;
 
+        this.$options.el = this.$options.el || el;
         this.$el = $(el);
 
         this._initProps();