+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Fix margin bottom behaviour in Grid Parallax
+- Fix error thrown by Notification component
 
 ## 3.0.0 beta 23 (May 18, 2017)